@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   '*.{js,ts,html,scss,md,json}': ['prettier --write'],
 };
